Fix training and exporting model for ffmpeg.
When training with Keras on tensorflow2, the `reset_after` default to True, as a result it generates twice as many bias variable for the GRU layer. `FFmpeg`'s rnnoise implementation cannot handle that properly.

See: https://stackoverflow.com/questions/57318930/calculating-the-number-of-parameters-of-a-gru-layer-keras

This commit explicitly set reset_after to False to make GRU consistent with tensorflow1 implementation. However, this also means we will not be able to leverage cuDNN on GPU to speed up the training.